pr hygiene
