+++ frontend/src/utils/helpers.ts
@@ -1,22 +1,6 @@
-import axios from 'axios'
-import type React from 'react'
-
-export const sendRequestThenSetState = async (
-  reactStateSetter: React.SetStateAction<any>,
-  abortController: AbortController,
-  httpMethod: string,
-  resource: string
-) => {
-  try {
-    const response = await axios({
-      method: httpMethod,
-      url: resource,
-      withCredentials: true,
-      signal: abortController.signal
-    });
-    reactStateSetter(response.data);
-  } catch (error) {
-    console.log(error)
-    return error
-  }
-};
+export const removeModal = () => {
+  const modalRoot = document.getElementById('modal') as HTMLElement;
+  modalRoot.style.display = 'none'
+  const html = document.querySelector('html');
+  (html as HTMLHtmlElement).style.overflow = 'auto';
+}
