--- frontend/src/utils/removeModal.ts
@@ -1,6 +1,0 @@
-export const removeModal = () => {
-  const modalRoot = document.getElementById('modal') as HTMLElement;
-  modalRoot.style.display = 'none'
-  const html = document.querySelector('html');
-  (html as HTMLHtmlElement).style.overflow = 'auto';
-}
